# docs/swagger.yaml
---
swagger: '2.0'
info:
  version: 1.0.0
  title: Bookmarks.dev api
  description: API supporting the bookmarks.dev website
  contact:
    name: Adrian Matei
    email: ama@codingpedia.org
    url: http://www.adrianmatei.me
  license:
    name: MIT
basePath: "/api"
paths:
  "/":
    get:
      description: message with link to Swagger docs
      produces:
        - text/html
      responses:
        '200':
          description: message with link to Swagger docs
  "/public/bookmarks":
    get:
      description: returns all public bookmarks
      produces:
        - application/json
      responses:
        '200':
          description: all public bookmarks
          schema:
            type: array
            items:
              "$ref": "#/definitions/bookmark"
  "/secured/public/bookmarks/{bookmarkId}":
    patch:
      description: updates rating of bookmark (a user starred or unstarred the bookmark)
      produces:
        - application/json
      responses:
        '200':
          description: OK when operation successful and updated bookmark in body
          schema:
            "$ref": "#/definitions/bookmark"
        '404':
          description: NOT_FOUND when bookmark cannot be found
          schema:
            "$ref": "#/definitions/myError"
        '401':
          description: unauthorized (you need to have a valid token) and the ratingUserId must be the same with the one in the request body
        '500':
          description: internal server error, when unknown server error is catched
          schema:
            "$ref": "#/definitions/myError"
  "/personal/users/{userId}/bookmarks":
    get:
      description: list of user's bookmarks
      operationId: getUserBookmarks
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: userId
          in: path
          description: keycloak user id
          required: true
          type: string
      responses:
        '200':
          description: list of user's bookmarks
          schema:
            type: array
            items:
              "$ref": "#/definitions/bookmark"
        '401':
          description: unauthorized (you need to have a valid token) and userId must match the subject in token
    post:
      description: create new bookmark for user
      operationId: createBookmark
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: userId
          in: path
          description: keycloak user id
          required: true
          type: string
        - in: body
          name: body
          description: Data that replaces placeholders in the sent emails
          required: true
          schema:
            "$ref": "#/definitions/bookmark"
      responses:
        '201':
          description: Empty response, status is relevant
          headers:
            location:
              schema:
                type: string
              description: resource url of the newly created bookmark
        '401':
          description: unauthorized (you need to have a valid token) and userId must match the subject in token
        '409':
          description: conflict

  "/personal/users/{userId}/bookmarks/{bookmarkId}":
    put:
      description: (full) update existing bookmark of user
      operationId: updateBookmark
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: userId
          in: path
          description: keycloak user id
          required: true
          type: string
        - name: bookmarkId
          in: path
          description: id of the bookmark
          required: true
          type: string
        - in: body
          name: body
          description: Complete bookmark to be updated
          required: true
          schema:
            "$ref": "#/definitions/bookmark"
      responses:
        '200':
          description: OK when operation successful and updated bookmark in body
          schema:
            "$ref": "#/definitions/bookmark"
        '404':
          description: NOT_FOUND when bookmark cannot be found for user
          schema:
            "$ref": "#/definitions/myError"
        '401':
          description: unauthorized (you need to have a valid token) and userId must match the subject in token
        '500':
          description: internal server error, when unknown server error is catched
          schema:
            "$ref": "#/definitions/myError"
    delete:
      description: deletes bookmark of user
      operationId: updateBookmark
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: userId
          in: path
          description: keycloak user id
          required: true
          type: string
        - name: bookmarkId
          in: path
          description: id of the bookmark
          required: true
          type: string
      responses:
        '204':
          description: OK (No content) when operation successful and bookmark was removed
        '404':
          description: NOT_FOUND when bookmark cannot be found for user
          schema:
            "$ref": "#/definitions/myError"
        '401':
          description: unauthorized (you need to have a valid token) and userId must match the subject in token
        '500':
          description: internal server error, when unknown server error is catched
          schema:
            "$ref": "#/definitions/myError"
definitions:
  bookmark:
    type: object
    required:
      - name
      - location
      - tags
      - userId
      - language
    properties:
      _id:
        type: string
      updatedAt:
        type: string
      createdAt:
        type: string
      description:
        type: string
      descriptionHtml:
        type: string
      language:
        type: string
      lastAccessedAt:
        type: string
      location:
        type: string
      name:
        type: string
      publishedOn:
        type: string
      public:
        type: boolean
      tags:
        type: array
        items:
          type: string
      userId:
        type: string
  myError:
    type: object
    properties:
      title:
        type: string
      messages:
        type: array
        items:
          type: string