# src/routes/users/bookmarks/personal-bookmarks.service.js
const Bookmark = require('../../../model/bookmark');
const User = require('../../../model/user');

const NotFoundError = require('../../../error/not-found.error');

const BookmarkInputValidator = require('../../../common/bookmark-input.validator');

/**
 * CREATE bookmark for user
 */
let createBookmark = async function (userId, bookmark) {
  BookmarkInputValidator.validateBookmarkInput(userId, bookmark);

  await BookmarkInputValidator.verifyPublicBookmarkExistenceOnCreation(bookmark);

  let newBookmark = await bookmark.save();

  return newBookmark;
}

/* GET suggested tags used for user */
let getSuggestedTagsForUser = async (userId) => {

  const tags = await Bookmark.distinct("tags",
    {
      $or: [
        {userId: userId},
        {public: true}
      ]
    }); // sort does not work with distinct in mongoose - https://mongoosejs.com/docs/api.html#query_Query-sort

  return tags;
};


/* GET tags used by user */
let getUserTags = async (userId) => {

  const tags = await Bookmark.distinct("tags",
    {userId: userId}
  ); // sort does not work with distinct in mongoose - https://mongoosejs.com/docs/api.html#query_Query-sort

  return tags;
};


/* GET bookmark of user by bookmarkId */
let getBookmarkById = async (userId, bookmarkId) => {

  const bookmark = await Bookmark.findOne({
    _id: bookmarkId,
    userId: userId
  });

  if (!bookmark) {
    throw new NotFoundError(`Bookmark NOT_FOUND the userId: ${userId} AND id: ${bookmarkId}`);
  } else {
    return bookmark;
  }
};

/* GET bookmark of user by location*/
let getBookmarkByLocation = async (userId, location) => {

  const bookmark = await Bookmark.findOne({
    userId: userId,
    location: location
  }).lean().exec();

  if (!bookmark) {
    throw new NotFoundError(`Bookmark NOT_FOUND the userId: ${userId} AND location: ${location}`);
  } else {
    return bookmark;
  }
};

/* GET last accessed bookmarks of the - currently there is a limit set to 100 */
let getLastAccessedBookmarks = async (userId) => {
  const bookmarks = await Bookmark.find({userId: userId})
    .sort({lastAccessedAt: -1})
    .limit(100);

  return bookmarks;
};

/* GET last created bookmarks of the user - currently there is a limit set to 100 */
let getLastCreatedBookmarks = async (userId) => {
  const bookmarks = await Bookmark.find({userId: userId})
    .sort({createdAt: -1})
    .limit(30);

  return bookmarks;
};

/* GET last created bookmarks of the user - currently there is a limit set to 100 */
let getMostLikedBookmarks = async (userId) => {
  const bookmarks = await Bookmark.find({userId: userId})
    .sort({likes: -1})
    .limit(30);

  return bookmarks;
};

/* GET last created bookmarks of the user - currently there is a limit set to 100 */
let getMostUsedBookmarks = async (userId) => {
  const bookmarks = await Bookmark.find({userId: userId, ownerVisitCount: {$exists: true}})
    .select('+ownerVisitCount')
    .sort({ownerVisitCount: -1})
    .limit(30);

  return bookmarks;
};

/**
 * full UPDATE via PUT - that is the whole document is required and will be updated
 * the descriptionHtml parameter is only set in backend, if only does not come front-end (might be an API call)
 */
let updateBookmark = async (userId, bookmarkId, bookmark) => {

  BookmarkInputValidator.validateBookmarkInput(userId, bookmark);

  await BookmarkInputValidator.verifyPublicBookmarkExistenceOnUpdate(bookmark, userId);

  const updatedBookmark = await Bookmark.findOneAndUpdate(
    {
      _id: bookmarkId,
      userId: userId
    },
    bookmark,
    {new: true}
  );

  const bookmarkNotFound = !updatedBookmark;
  if (bookmarkNotFound) {
    throw new NotFoundError('Bookmark NOT_FOUND with id: ' + bookmarkId + ' AND location: ' + bookmark.location);
  } else {
    return updatedBookmark;
  }
};

let increaseOwnerVisitCount = async (userId, bookmarkId) => {
  await Bookmark.findOneAndUpdate(
    {
      _id: bookmarkId,
      userId: userId
    },
    {
      $inc: {ownerVisitCount: 1}
    }
  );
}


/*
* DELETE bookmark for user
*/
let deleteBookmarkById = async (userId, bookmarkId) => {
  const bookmark = await Bookmark.findOneAndRemove({
    _id: bookmarkId,
    userId: userId
  });

  if (!bookmark) {
    throw new NotFoundError('Bookmark NOT_FOUND with id: ' + bookmarkId);
  } else {
    await User.update(
      {},
      {
        $pull: {
          readLater: bookmarkId,
          likes: bookmarkId,
          pinned: bookmarkId,
          history: bookmarkId,
          favorites: bookmarkId
        }
      },
      {multi: true}
    );

    return true;
  }
};

/*
* DELETE bookmark for user by "location"
*/
let deleteBookmarkByLocation = async (userId, location) => {
  const bookmark = await Bookmark.findOneAndRemove({
    location: location,
    userId: userId
  });

  if (!bookmark) {
    throw new NotFoundError(`Bookmark NOT_FOUND the userId: ${userId} AND location: ${location}`);
  }

  return true;
};

/*
* DELETE private bookmarks for user by "tag"
*/
let deletePrivateBookmarksByTag = async (userId, tag) => {
  await Bookmark.deleteMany({
    userId: userId,
    public: false,
    tags: tag
  });

  return true;
};

module.exports = {
  createBookmark: createBookmark,
  getSuggestedTagsForUser: getSuggestedTagsForUser,
  getBookmarkById: getBookmarkById,
  getBookmarkByLocation: getBookmarkByLocation,
  getLastAccessedBookmarks: getLastAccessedBookmarks,
  getLastCreatedBookmarks: getLastCreatedBookmarks,
  getMostUsedBookmarks: getMostUsedBookmarks,
  getMostLikedBookmarks: getMostLikedBookmarks,
  updateBookmark: updateBookmark,
  increaseOwnerVisitCount: increaseOwnerVisitCount,
  deleteBookmarkById: deleteBookmarkById,
  deleteBookmarkByLocation: deleteBookmarkByLocation,
  deletePrivateBookmarksByTag: deletePrivateBookmarksByTag,
  getUserTags: getUserTags
};